chore: initialize frontend project structure
- Migrated from react-scripts to vite build tool
- Migrated to Typescript
- ESLint, Prettier + Husky
- ignore build (dist) directory
