+++ .gitignore
@@ -1,4 +1,6 @@
 frontend/node_modules
+frontend/dist/*
+
 backend/.idea/*
 backend/*.iml
 backend/target/**
