+++ frontend/src/App.tsx
@@ -1,0 +1,31 @@
+import React, { Component } from 'react';
+//import logo from './logo.svg';
+
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <header className="App-header">
+          <h2>Welcome to the interview app!</h2>
+          <p>
+            Edit <code>src/App.js</code> and save to reload.
+          </p>
+
+          <ol>
+            <li>
+              Fetch Data from a public API{' '}
+              <a href="https://github.com/toddmotto/public-apis">Samples</a>
+            </li>
+            <li>Display data from API onto your page (Table, List, etc.)</li>
+            <li>
+              Apply a styling solution of your choice to make your page look
+              different (CSS, SASS, CSS-in-JS)
+            </li>
+          </ol>
+        </header>
+      </div>
+    );
+  }
+}
+
+export default App;
